refactor(langgraph-starter): move catalog to app/declarative-generative-ui/
Move demonstration catalog code from src/a2ui/ to
src/app/declarative-generative-ui/ with cleaner file names:

  definitions.ts  — Zod schemas (server-safe, imported by route.ts)
  renderers.tsx   — React components (imported by layout.tsx)

Remove old src/a2ui/ directory entirely (theme.css was unused).
