+++ examples/integrations/langgraph-python/apps/app/src/app/declarative-generative-ui/definitions.ts
